pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "omagent"
version = "0.1.0"
description = "Generic domain-configurable agentic engine with interactive CLI and server mode"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.40.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "websockets>=12.0",
    "prompt_toolkit>=3.0.43",
    "rich>=13.7.0",
    "click>=8.1.7",
    "aiosqlite>=0.20.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "jupyter-client>=8.6.0",
    "ipykernel>=6.29.0",
    "pandas>=2.2.0",
    "matplotlib>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
]

[project.scripts]
omagent = "omagent.cli.main:cli"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
